♻️ Refactor: index 훅 로직 분리
